Move to Shared for SqlDependency.cs (#1299)
